# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Gacha Game Backend API
  version: 1.0.0
  description: API for managing gacha rolls, user inventories, auctions, transactions, and user authentication.
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /users/auth/login:
    post:
      tags:
        - PLAYER
      summary: Login a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: player123
                password:
                  type: string
                  example: securepassword
      responses:
        '200':
          description: Successful login
        '401':
          description: Invalid credentials
        '500':
          description: Database error
        '501':
          description: Unexpected error
  /users/auth/register:
    post:
      tags:
        - PLAYER
      summary: Register a new player
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: new_player123
                password:
                  type: string
                  example: securepassword
                email:
                  type: string
                  example: new_player@example.com
                image:
                  type: object
      responses:
        '200':
          description: Player successfully registered
        '400':
          description: Missing required fields
        '409':
          description: Username or email already exists
        '500':
          description: Unexpected error
  /users/auth/logout:
    delete:
      tags:
        - PLAYER
      summary: Logout a player
      responses:
        '200':
          description: Successful logout
        '401':
          description: Invalid user type
        '500':
          description: Unexpected error        
...